Import vst2sdk lib
